Fix bugs with formula_indexer

+++ search/formula_indexer.py
@@ -7,41 +7,51 @@ import features_extractor as fe
 import re
 
 
-def reindex_all_formulas():
+def reindex_all_formulas(reset_formula=False):
     """
-    Drops and creates formula and formula index table.
+    Reindex the formula and formula index table.
+
+    Args:
+        reset_formula: Option to drop the formula index table and recreates it.
     """
     questions = Question.objects.all()
 
-    Formula.objects.all().delete()
+    if reset_formula:
+        Formula.objects.all().delete()
+    else:
+        Formula.objects.all().update(status=False)
+
     FormulaIndex.objects.all().delete()
 
     # Reindex formulas in every question 
     for question in questions:
-        reindex_formulas_in_question(question.id)
+        reindex_formulas_in_question(question.id, reset_formula)
 
 
-def reindex_formulas_in_question(question_id):
+def reindex_formulas_in_question(question_id, create_formula=False):
     """
     Creates formula and formula index table from a question.
 
     Args:
+        create_formula: Option to to create new formulas from the question.
         question_id: question id
     """
     question = Question.objects.get(id=question_id)
-    formulas = extract_formulas_from_question(question_id)
 
-    print("Question #%d generates %d formulas" % (question_id, len(formulas)))
+    if create_formula:
+        formulas = extract_formulas_from_question(question_id)
 
-    for formula in formulas:
-        new_formula = Formula(content=formula, status=False, question=question)
-        new_formula.save()
+        for formula_str in formulas:
+            new_formula = Formula(content=formula_str, status=False,
+                                  question=question)
+            new_formula.save()
 
-        try:
-            formula_id = new_formula.id
-            create_formula_index_model(formula, formula_id)
-        except (KeyError, Formula.DoesNotExist):
-            print "Error"
+    formulas = question.formula_set.all()
+    try:
+        for formula in formulas:
+            create_formula_index_model(formula.content, formula.id)
+    except (KeyError, Formula.DoesNotExist):
+        print "Could not create formula index."
 
 
 def extract_formulas_from_question(question_id):
@@ -93,14 +103,14 @@ def create_formula_index_model(latex_str, formula_id):
             formula_obj.save()
 
             # Create index term in FormulaIndex table
-            for term in chain(inorder_sem_terms, sorted_sem_terms):
+            for term in chain.from_iterable(inorder_sem_terms+sorted_sem_terms):
                 create_update_formula_index(formula_obj.id, term)
 
             for term in chain(struc_features, const_features, var_features):
                 create_update_formula_index(formula_obj.id, term)
 
-    except (KeyError, Formula.DoesNotExist):
-        print "Error"
+    except (KeyError, Formula.DoesNotExist) as err:
+        print err
 
 
 def create_update_formula_index(formula_id, term):
@@ -112,6 +122,7 @@ def create_update_formula_index(formula_id, term):
         formula_id: formula id.
         term: the formula feature.
     """
+
     try:
         f_index = FormulaIndex.objects.get(pk=term)
         formulaid_str = str(formula_id)
